Move updater event args to a separate namespace
